refactor: user - redisUtil 분리#149
