+++ user-service/src/main/java/com/boeingmerryho/business/userservice/application/UserVerificationHelper.java
@@ -1,0 +1,18 @@
+package com.boeingmerryho.business.userservice.application;
+
+public interface UserVerificationHelper {
+
+	String generateVerificationCode();
+
+	void storeVerificationCode(String email, String code);
+
+	String getVerificationCode(String email);
+
+	void removeVerificationCode(String email);
+
+	void checkDuplicatedVerificationRequest(String email);
+
+	String getNotifyLoginResponse(Long id);
+
+}
+
