+++ user-service/src/main/java/com/boeingmerryho/business/userservice/application/service/UserAdminService.java
@@ -12,6 +12,7 @@ import org.springframework.transaction.annotation.Transactional;
 
 import com.boeingmerryho.business.userservice.application.UserHelper;
 import com.boeingmerryho.business.userservice.application.UserJwtHelper;
+import com.boeingmerryho.business.userservice.application.UserVerificationHelper;
 import com.boeingmerryho.business.userservice.application.dto.mapper.UserApplicationMapper;
 import com.boeingmerryho.business.userservice.application.dto.request.admin.UserAdminCheckEmailRequestServiceDto;
 import com.boeingmerryho.business.userservice.application.dto.request.admin.UserAdminDeleteRequestServiceDto;
@@ -30,6 +31,7 @@ import com.boeingmerryho.business.userservice.application.dto.request.other.User
 import com.boeingmerryho.business.userservice.application.dto.response.admin.UserAdminFindResponseDto;
 import com.boeingmerryho.business.userservice.application.dto.response.inner.UserTokenResult;
 import com.boeingmerryho.business.userservice.application.dto.response.other.UserLoginResponseServiceDto;
+import com.boeingmerryho.business.userservice.application.utils.RedisUtil;
 import com.boeingmerryho.business.userservice.application.utils.mail.EmailService;
 import com.boeingmerryho.business.userservice.domain.User;
 import com.boeingmerryho.business.userservice.domain.UserRoleType;
@@ -60,7 +62,9 @@ public class UserAdminService {
 	private final PasswordEncoder passwordEncoder;
 
 	private final UserHelper userHelper;
+	private final RedisUtil redisUtil;
 	private final UserJwtHelper userJwtHelper;
+	private final UserVerificationHelper userVerificationHelper;
 	private final EmailService emailService;
 
 	@Value("${admin.key}")
@@ -73,7 +77,7 @@ public class UserAdminService {
 
 		User user = User.withAdminRole(
 			dto.username(),
-			userHelper.encodePassword(dto.password(), passwordEncoder),
+			userHelper.encodePassword(dto.password()),
 			dto.email(),
 			dto.nickname(),
 			dto.birth()
@@ -125,7 +129,7 @@ public class UserAdminService {
 
 		UserRoleType oldRole = user.getRole();
 		user.updateRoleType(dto.newRole());
-		userHelper.updateRedisUserInfo(user);
+		redisUtil.updateUserInfo(user);
 
 		return userApplicationMapper.toUserAdminUpdateRoleResponseDto(dto.id(), oldRole, dto.newRole());
 	}
@@ -135,7 +139,7 @@ public class UserAdminService {
 		User user = userHelper.findUserById(dto.id());
 
 		user.deleteRoleType();
-		userHelper.updateRedisUserInfo(user);
+		redisUtil.updateUserInfo(user);
 		return user.getId();
 	}
 
@@ -144,7 +148,7 @@ public class UserAdminService {
 		User user = userHelper.findUserById(dto.id());
 		user.softDelete(user.getId());
 
-		userHelper.clearRedisUserData(user.getId());
+		redisUtil.clearRedisUserData(user.getId());
 		return user.getId();
 	}
 
@@ -153,24 +157,24 @@ public class UserAdminService {
 
 		UserTokenResult result = userHelper.getUserTokenFromRedis(userId);
 		String accessToken = (String)result.token().get("accessToken");
-		userHelper.blacklistToken(accessToken);
+		redisUtil.blacklistToken(accessToken);
 
-		userHelper.deleteFromRedisByKey(result.tokenKey());
+		redisUtil.deleteFromRedisByKey(result.tokenKey());
 
-		userHelper.clearRedisUserData(userId);
+		redisUtil.clearRedisUserData(userId);
 	}
 
 	public UserLoginResponseDto loginUserAdmin(UserAdminLoginRequestServiceDto dto) {
 		User user = userHelper.findUserByEmail(dto.email());
-		userHelper.updateRedisUserInfo(user);
+		redisUtil.updateUserInfo(user);
 
-		Map<String, String> tokenMap = userHelper.updateUserJwtTokenRedis(user.getId());
+		Map<String, String> tokenMap = redisUtil.updateUserJwtToken(user.getId());
 		UserLoginResponseServiceDto serviceDto = UserLoginResponseServiceDto.fromTokens(
 			tokenMap.get("accessToken"),
 			tokenMap.get("refreshToken")
 		);
 
-		userHelper.getNotifyLoginResponse(user.getId());
+		userVerificationHelper.getNotifyLoginResponse(user.getId());
 
 		return userApplicationMapper.toUserLoginResponseDto(serviceDto);
 	}
@@ -180,7 +184,7 @@ public class UserAdminService {
 		User user = userHelper.findUserById(dto.id());
 		user.softDelete(user.getId());
 
-		userHelper.clearRedisUserData(user.getId());
+		redisUtil.clearRedisUserData(user.getId());
 		return user.getId();
 	}
 
@@ -193,7 +197,7 @@ public class UserAdminService {
 
 		Long userId = userJwtHelper.getUserIdFromToken(refreshToken);
 
-		userHelper.isEqualStoredRefreshToken(userId, refreshToken);
+		redisUtil.isEqualStoredRefreshToken(userId, refreshToken);
 
 		String newAccessToken = userJwtHelper.generateAccessToken(userId);
 
@@ -228,7 +232,7 @@ public class UserAdminService {
 		User user = userHelper.findUserById(dto.id());
 
 		if (!userHelper.isEmpty(dto.password())) {
-			String encodedPassword = userHelper.encodePassword(dto.password(), passwordEncoder);
+			String encodedPassword = userHelper.encodePassword(dto.password());
 			user.updatePassword(encodedPassword);
 		}
 		if (!userHelper.isEmpty(dto.username())) {
@@ -241,28 +245,28 @@ public class UserAdminService {
 			user.updateBirth(dto.birth());
 		}
 
-		userHelper.updateRedisUserInfo(user);
+		redisUtil.updateUserInfo(user);
 
 		return userApplicationMapper.toUserAdminUpdateResponseDto(user.getId());
 	}
 
 	public UserAdminVerificationResponseDto sendVerificationCode(
 		UserAdminEmailVerificationRequestServiceDto dto) {
-		userHelper.checkDuplicatedVerificationRequest(dto.email());
+		userVerificationHelper.checkDuplicatedVerificationRequest(dto.email());
 		userHelper.verifyEmailFormat(dto.email());
-		String code = userHelper.generateVerificationCode();
-		userHelper.storeVerificationCode(dto.email(), code);
+		String code = userVerificationHelper.generateVerificationCode();
+		userVerificationHelper.storeVerificationCode(dto.email(), code);
 		emailService.sendVerificationEmail(dto.email(), code);
 		return userApplicationMapper.toUserAdminVerificationResponseDto(dto.email()).success("인증 메일 발송 성공");
 	}
 
 	public UserAdminVerificationResponseDto verifyCode(
 		UserAdminEmailVerificationCheckRequestServiceDto dto) {
-		String storedCode = userHelper.getVerificationCode(dto.email());
+		String storedCode = userVerificationHelper.getVerificationCode(dto.email());
 		if (storedCode == null || !storedCode.equals(dto.code())) {
 			throw new GlobalException(ErrorCode.VERIFICATION_FAIL);
 		}
-		userHelper.removeVerificationCode(dto.email());
+		userVerificationHelper.removeVerificationCode(dto.email());
 
 		return UserAdminVerificationResponseDto.success("메일 인증 성공");
 	}
