+++ user-service/src/main/java/com/boeingmerryho/business/userservice/application/service/UserService.java
@@ -3,13 +3,12 @@ package com.boeingmerryho.business.userservice.application.service;
 import java.util.Map;
 
 import org.springframework.cache.annotation.Cacheable;
-import org.springframework.data.redis.core.RedisTemplate;
-import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import com.boeingmerryho.business.userservice.application.UserHelper;
 import com.boeingmerryho.business.userservice.application.UserJwtHelper;
+import com.boeingmerryho.business.userservice.application.UserVerificationHelper;
 import com.boeingmerryho.business.userservice.application.dto.mapper.UserApplicationMapper;
 import com.boeingmerryho.business.userservice.application.dto.request.other.UserCheckEmailRequestServiceDto;
 import com.boeingmerryho.business.userservice.application.dto.request.other.UserEmailVerificationCheckRequestServiceDto;
@@ -23,6 +22,7 @@ import com.boeingmerryho.business.userservice.application.dto.request.other.User
 import com.boeingmerryho.business.userservice.application.dto.request.other.UserWithdrawRequestServiceDto;
 import com.boeingmerryho.business.userservice.application.dto.response.inner.UserTokenResult;
 import com.boeingmerryho.business.userservice.application.dto.response.other.UserLoginResponseServiceDto;
+import com.boeingmerryho.business.userservice.application.utils.RedisUtil;
 import com.boeingmerryho.business.userservice.application.utils.mail.EmailService;
 import com.boeingmerryho.business.userservice.domain.User;
 import com.boeingmerryho.business.userservice.domain.repository.UserRepository;
@@ -45,10 +45,10 @@ public class UserService {
 
 	private final UserRepository userRepository;
 	private final UserApplicationMapper userApplicationMapper;
-	private final PasswordEncoder passwordEncoder;
-	private final RedisTemplate<String, Object> redisTemplate;
 	private final UserHelper userHelper;
 	private final UserJwtHelper userJwtHelper;
+	private final UserVerificationHelper userVerificationHelper;
+	private final RedisUtil redisUtil;
 	private final EmailService emailService;
 
 	@Transactional
@@ -57,7 +57,7 @@ public class UserService {
 
 		User user = User.withDefaultRole(
 			dto.username(),
-			userHelper.encodePassword(dto.password(), passwordEncoder),
+			userHelper.encodePassword(dto.password()),
 			dto.email(),
 			dto.nickname(),
 			dto.birth()
@@ -68,15 +68,15 @@ public class UserService {
 
 	public UserLoginResponseDto loginUser(UserLoginRequestServiceDto dto) {
 		User user = userHelper.findUserByEmail(dto.email());
-		userHelper.updateRedisUserInfo(user);
+		redisUtil.updateUserInfo(user);
 
-		Map<String, String> tokenMap = userHelper.updateUserJwtTokenRedis(user.getId());
+		Map<String, String> tokenMap = redisUtil.updateUserJwtToken(user.getId());
 		UserLoginResponseServiceDto serviceDto = UserLoginResponseServiceDto.fromTokens(
 			tokenMap.get("accessToken"),
 			tokenMap.get("refreshToken")
 		);
 
-		userHelper.getNotifyLoginResponse(user.getId());
+		userVerificationHelper.getNotifyLoginResponse(user.getId());
 
 		return userApplicationMapper.toUserLoginResponseDto(serviceDto);
 	}
@@ -86,11 +86,11 @@ public class UserService {
 
 		UserTokenResult result = userHelper.getUserTokenFromRedis(userId);
 		String accessToken = (String)result.token().get("accessToken");
-		userHelper.blacklistToken(accessToken);
+		redisUtil.blacklistToken(accessToken);
 
-		userHelper.deleteFromRedisByKey(result.tokenKey());
+		redisUtil.deleteFromRedisByKey(result.tokenKey());
 
-		userHelper.clearRedisUserData(userId);
+		redisUtil.clearRedisUserData(userId);
 	}
 
 	@Transactional(readOnly = true)
@@ -106,7 +106,7 @@ public class UserService {
 		User user = userHelper.findUserById(dto.id());
 
 		if (!userHelper.isEmpty(dto.password())) {
-			String encodedPassword = userHelper.encodePassword(dto.password(), passwordEncoder);
+			String encodedPassword = userHelper.encodePassword(dto.password());
 			user.updatePassword(encodedPassword);
 		}
 		if (!userHelper.isEmpty(dto.username())) {
@@ -119,7 +119,7 @@ public class UserService {
 			user.updateBirth(dto.birth());
 		}
 
-		userHelper.updateRedisUserInfo(user);
+		redisUtil.updateUserInfo(user);
 
 		return userApplicationMapper.toUserAdminUpdateResponseDto(user.getId());
 	}
@@ -129,7 +129,7 @@ public class UserService {
 		User user = userHelper.findUserById(dto.id());
 		user.softDelete(user.getId());
 
-		userHelper.clearRedisUserData(user.getId());
+		redisUtil.clearRedisUserData(user.getId());
 
 		return user.getId();
 	}
@@ -149,7 +149,7 @@ public class UserService {
 
 		Long userId = userJwtHelper.getUserIdFromToken(refreshToken);
 
-		userHelper.isEqualStoredRefreshToken(userId, refreshToken);
+		redisUtil.isEqualStoredRefreshToken(userId, refreshToken);
 
 		String newAccessToken = userJwtHelper.generateAccessToken(userId);
 
@@ -158,20 +158,20 @@ public class UserService {
 
 	public UserVerificationResponseDto sendVerificationCode(
 		UserEmailVerificationRequestServiceDto dto) {
-		userHelper.checkDuplicatedVerificationRequest(dto.email());
+		userVerificationHelper.checkDuplicatedVerificationRequest(dto.email());
 		userHelper.verifyEmailFormat(dto.email());
-		String code = userHelper.generateVerificationCode();
-		userHelper.storeVerificationCode(dto.email(), code);
+		String code = userVerificationHelper.generateVerificationCode();
+		userVerificationHelper.storeVerificationCode(dto.email(), code);
 		emailService.sendVerificationEmail(dto.email(), code);
 		return userApplicationMapper.toUserVerificationResponseDto(dto.email()).success("인증 메일 발송 성공");
 	}
 
 	public UserVerificationResponseDto verifyCode(UserEmailVerificationCheckRequestServiceDto dto) {
-		String storedCode = userHelper.getVerificationCode(dto.email());
+		String storedCode = userVerificationHelper.getVerificationCode(dto.email());
 		if (storedCode == null || !storedCode.equals(dto.code())) {
 			throw new GlobalException(ErrorCode.VERIFICATION_FAIL);
 		}
-		userHelper.removeVerificationCode(dto.email());
+		userVerificationHelper.removeVerificationCode(dto.email());
 
 		return UserVerificationResponseDto.success("메일 인증 성공");
 	}
